refactor(cmd): Extract quota helper in utils
The hmsl quota command should not import code from the quota command.
This is needed to enforce boundaries between module as planned in #521.

refs #521.
